--- a/decoder/stream_decoder_test.go
+++ b/decoder/stream_decoder_test.go
@@ -18,6 +18,7 @@
 package decoder
 
 import (
+	"bufio"
 	"bytes"
 	"strings"
 	"testing"
@@ -61,7 +62,7 @@ func TestNDStreamReader(t *testing.T) {
 		},
 	}
 	buf := bytes.NewBufferString(strings.Join(lines, "\n"))
-	n := NewNDJSONStreamReader(buf, 20)
+	n := NewNDJSONStreamReader(NewLineReader(bufio.NewReaderSize(buf, 20), 20))
 
 	for idx, test := range expected {
 		out, err := n.Read()