--- a/processor/stream/benchmark_test.go
+++ b/processor/stream/benchmark_test.go
@@ -22,12 +22,9 @@ import (
 	"context"
 	"errors"
 	"io/ioutil"
-	"math"
 	"path/filepath"
 	"runtime"
 	"testing"
-
-	r "golang.org/x/time/rate"
 
 	"github.com/elastic/apm-server/publish"
 	"github.com/elastic/apm-server/tests/loader"
@@ -47,8 +44,8 @@ func BenchmarkStreamProcessor(b *testing.B) {
 		b.Error(err)
 	}
 	//ensure to not hit rate limit as blocking wait would be measured otherwise
-	rl := r.NewLimiter(r.Limit(math.MaxFloat64-1), math.MaxInt32)
-	sp := &StreamProcessor{MaxEventSize: 100000}
+	// rl := r.NewLimiter(r.Limit(math.MaxFloat64-1), math.MaxInt32)
+	sp := &StreamProcessor{MaxEventSize: 300 * 1024}
 	for _, f := range files {
 		b.Run(f.Name(), func(b *testing.B) {
 			data, err := loader.LoadDataAsBytes(filepath.Join(dir, f.Name()))
@@ -63,7 +60,7 @@ func BenchmarkStreamProcessor(b *testing.B) {
 				b.StopTimer()
 				r.Reset(data)
 				b.StartTimer()
-				sp.HandleStream(context.Background(), rl, map[string]interface{}{}, ioutil.NopCloser(r), report)
+				sp.HandleStream(context.Background(), nil, map[string]interface{}{}, r, report)
 			}
 		})
 	}