--- a/processor/stream/package_tests/metadata_attrs_test.go
+++ b/processor/stream/package_tests/metadata_attrs_test.go
@@ -18,6 +18,7 @@
 package package_tests
 
 import (
+	"bufio"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -82,7 +83,8 @@ func getMetadataEventAttrs(t *testing.T, prefix string) *tests.Set {
 	payloadStream, err := loader.LoadDataAsStream("../testdata/intake-v2/only-metadata.ndjson")
 	require.NoError(t, err)
 
-	metadata, err := decoder.NewNDJSONStreamReader(payloadStream, 100*1024).Read()
+	lr := decoder.NewLineReader(bufio.NewReader(payloadStream), lrSize)
+	metadata, err := decoder.NewNDJSONStreamReader(lr).Read()
 	require.NoError(t, err)
 
 	contextMetadata := metadata["metadata"]