After our call a couple of days ago where it was pointed out how the 300kb buffer for the line reader gets allocated up front I had a sort of nightmare about the overhead per http request to the APM Server and wanted to look into if it was worth to reuse the buffers between requests. For large requests, the cost of the allocation is not significant, but for many small requests there is a speedup. 

While experimenting, I took the rate limiting out of the benchmark and changed it to use the 300kb default line length, as you can see below.  Rate limiting is only in effect in RUM, so the number here would correspond to the data from backend agents. 

This change makes the code more complex and I expect the overhead in beats for JSON serializing etc. (which are not measured in this benchmark) might make the gains here less attractive, but it's worth a discussion. 

**ns/op:**
```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkStreamProcessor/errors.ndjson-4                    280910        200638        -28.58%
BenchmarkStreamProcessor/events.ndjson-4                    146007        88538         -39.36%
BenchmarkStreamProcessor/heavy.ndjson-4                     24615514      25014177      +1.62%
```

**MB/s:**
```
benchmark                                                   old MB/s     new MB/s     speedup
BenchmarkStreamProcessor/errors.ndjson-4                    15.67        21.94        1.40x
BenchmarkStreamProcessor/events.ndjson-4                    6.90         11.37        1.65x
BenchmarkStreamProcessor/heavy.ndjson-4                     16.19        15.93        0.98x
```



<details>

```
$ benchcmp old.txt new.txt
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkStreamProcessor/errors.ndjson-4                    280910        200638        -28.58%
BenchmarkStreamProcessor/events.ndjson-4                    146007        88538         -39.36%
BenchmarkStreamProcessor/heavy.ndjson-4                     24615514      25014177      +1.62%
BenchmarkStreamProcessor/invalid-event.ndjson-4             121174        61915         -48.90%
BenchmarkStreamProcessor/invalid-json-event.ndjson-4        98437         43323         -55.99%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson-4     46482         7196          -84.52%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson-4        44754         6398          -85.70%
BenchmarkStreamProcessor/invalid-metadata.ndjson-4          52331         11085         -78.82%
BenchmarkStreamProcessor/metadata-null-values.ndjson-4      101302        41591         -58.94%
BenchmarkStreamProcessor/metricsets.ndjson-4                131967        74594         -43.48%
BenchmarkStreamProcessor/minimal-service.ndjson-4           98320         43179         -56.08%
BenchmarkStreamProcessor/only-metadata.ndjson-4             89040         39160         -56.02%
BenchmarkStreamProcessor/optional-timestamps.ndjson-4       138138        83302         -39.70%
BenchmarkStreamProcessor/ratelimit.ndjson-4                 374608        296282        -20.91%
BenchmarkStreamProcessor/spans.ndjson-4                     221895        160157        -27.82%
BenchmarkStreamProcessor/transactions.ndjson-4              214398        158546        -26.05%
BenchmarkStreamProcessor/unrecognized-event.ndjson-4        75768         29123         -61.56%

benchmark                                                   old MB/s     new MB/s     speedup
BenchmarkStreamProcessor/errors.ndjson-4                    15.67        21.94        1.40x
BenchmarkStreamProcessor/events.ndjson-4                    6.90         11.37        1.65x
BenchmarkStreamProcessor/heavy.ndjson-4                     16.19        15.93        0.98x
BenchmarkStreamProcessor/invalid-event.ndjson-4             6.32         12.37        1.96x
BenchmarkStreamProcessor/invalid-json-event.ndjson-4        5.96         13.55        2.27x
BenchmarkStreamProcessor/invalid-json-metadata.ndjson-4     9.85         63.64        6.46x
BenchmarkStreamProcessor/invalid-metadata-2.ndjson-4        10.01        70.01        6.99x
BenchmarkStreamProcessor/invalid-metadata.ndjson-4          8.73         41.23        4.72x
BenchmarkStreamProcessor/metadata-null-values.ndjson-4      4.91         11.95        2.43x
BenchmarkStreamProcessor/metricsets.ndjson-4                6.71         11.88        1.77x
BenchmarkStreamProcessor/minimal-service.ndjson-4           4.32         9.84         2.28x
BenchmarkStreamProcessor/only-metadata.ndjson-4             6.28         14.27        2.27x
BenchmarkStreamProcessor/optional-timestamps.ndjson-4       7.43         12.33        1.66x
BenchmarkStreamProcessor/ratelimit.ndjson-4                 11.24        14.22        1.27x
BenchmarkStreamProcessor/spans.ndjson-4                     11.97        16.58        1.39x
BenchmarkStreamProcessor/transactions.ndjson-4              13.18        17.82        1.35x
BenchmarkStreamProcessor/unrecognized-event.ndjson-4        5.16         13.43        2.60x

benchmark                                                   old allocs     new allocs     delta
BenchmarkStreamProcessor/errors.ndjson-4                    946            943            -0.32%
BenchmarkStreamProcessor/events.ndjson-4                    377            374            -0.80%
BenchmarkStreamProcessor/heavy.ndjson-4                     134559         134556         -0.00%
BenchmarkStreamProcessor/invalid-event.ndjson-4             270            266            -1.48%
BenchmarkStreamProcessor/invalid-json-event.ndjson-4        199            196            -1.51%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson-4     29             27             -6.90%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson-4        24             22             -8.33%
BenchmarkStreamProcessor/invalid-metadata.ndjson-4          52             49             -5.77%
BenchmarkStreamProcessor/metadata-null-values.ndjson-4      178            175            -1.69%
BenchmarkStreamProcessor/metricsets.ndjson-4                296            293            -1.01%
BenchmarkStreamProcessor/minimal-service.ndjson-4           201            198            -1.49%
BenchmarkStreamProcessor/only-metadata.ndjson-4             212            209            -1.42%
BenchmarkStreamProcessor/optional-timestamps.ndjson-4       388            385            -0.77%
BenchmarkStreamProcessor/ratelimit.ndjson-4                 1448           1444           -0.28%
BenchmarkStreamProcessor/spans.ndjson-4                     767            764            -0.39%
BenchmarkStreamProcessor/transactions.ndjson-4              657            654            -0.46%
BenchmarkStreamProcessor/unrecognized-event.ndjson-4        128            126            -1.56%

benchmark                                                   old bytes     new bytes     delta
BenchmarkStreamProcessor/errors.ndjson-4                    361917        55556         -84.65%
BenchmarkStreamProcessor/events.ndjson-4                    334515        25106         -92.49%
BenchmarkStreamProcessor/heavy.ndjson-4                     7744681       7452358       -3.77%
BenchmarkStreamProcessor/invalid-event.ndjson-4             327101        17011         -94.80%
BenchmarkStreamProcessor/invalid-json-event.ndjson-4        324075        13841         -95.73%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson-4     313344        2157          -99.31%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson-4        313184        1987          -99.37%
BenchmarkStreamProcessor/invalid-metadata.ndjson-4          314585        3320          -98.94%
BenchmarkStreamProcessor/metadata-null-values.ndjson-4      322977        12370         -96.17%
BenchmarkStreamProcessor/metricsets.ndjson-4                330364        20481         -93.80%
BenchmarkStreamProcessor/minimal-service.ndjson-4           325331        15142         -95.35%
BenchmarkStreamProcessor/only-metadata.ndjson-4             325320        15047         -95.37%
BenchmarkStreamProcessor/optional-timestamps.ndjson-4       334785        25205         -92.47%
BenchmarkStreamProcessor/ratelimit.ndjson-4                 393111        86351         -78.03%
BenchmarkStreamProcessor/spans.ndjson-4                     353608        46186         -86.94%
BenchmarkStreamProcessor/transactions.ndjson-4              349446        41338         -88.17%
BenchmarkStreamProcessor/unrecognized-event.ndjson-4        321277        10633         -96.69%
```

</details>

Builds on https://github.com/elastic/apm-server/pull/1473