@roncohen 2018-11-01
Proper PR is here: https://github.com/elastic/apm-server/pull/1505